backend/pyproject.toml
[project]
name = "fecot-backend"
version = "0.1.0"
description = "Backend da plataforma FECOT - Federação Centro-Oeste de Taekwondo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Database
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",  # pinned — 4.1+ quebra compatibilidade com passlib atual

    # Validação CPF/CNPJ (opcional, mas útil)
    "validate-docbr>=1.10.0",

    # Utils
    "python-multipart>=0.0.12",  # uploads (pra futuro)
    "python-dotenv>=1.0.1",
    "email-validator>=2.2.0",
    "Pillow>=10.4.0",            # processamento de imagens (avatar)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # linha longa tratada pelo formatter

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
